Fix bugs in Program and renderer.

+++ native/cocos/renderer/renderer/Light.h
@@ -198,11 +198,11 @@ public:
     /**
      *  @brief Sets the shadow darkness setting.
      */
-    inline void setShadowDarkness(uint32_t val) { _shadowDarkness = val; }
+    inline void setShadowDarkness(float val) { _shadowDarkness = val; }
     /**
      *  @brief Gets the shadow darkness setting.
      */
-    inline uint32_t getShadowDarkness() const { return _shadowDarkness; }
+    inline float getShadowDarkness() const { return _shadowDarkness; }
     
     /**
      *  @brief Sets the min depth of shadow.
@@ -255,7 +255,7 @@ public:
     /**
      *  @brief Sets the shadow frustum size.
      */
-    inline void setShadowFrustumSize(uint32_t val) {_shadowFustumSize = val;};
+    inline void setShadowFrustumSize(float val) {_shadowFustumSize = val;};
     
 private:
     void updateLightPositionAndDirection();
@@ -295,7 +295,7 @@ private:
     uint32_t _frustumEdgeFalloff = 0; // used by directional and spot light.
     Mat4 _viewProjMatrix;
     float _spotAngleScale = 1; // used for spot light.
-    uint32_t _shadowFustumSize = 50; // used for directional light.
+    float _shadowFustumSize = 50; // used for directional light.
     
     Mat4 _worldMatrix;
     Mat4 _worldRT;
